requirements.txt
torchjd
torchio
zarr
pandas
numba